% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{beta_ineq}
\alias{beta_ineq}
\title{Calculate Pr(X > Y + delta) where X and Y are independent Beta random variables
using numerical integration}
\usage{
beta_ineq(a, b, c, d, delta = 0, ...)
}
\arguments{
\item{delta}{The difference we wish to assess (i.e. X - Y > delta)}

\item{...}{other arguments passed to integrate/quadgk function}
}
\value{
The value of the integral
}
\description{
Calculate Pr(X > Y + delta) where X and Y are independent Beta random variables
using numerical integration
}
