% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{beta_ineq_sim}
\alias{beta_ineq_sim}
\title{Calculate Pr(X > Y + delta) where X and Y are independent Beta random variables
using Monte Carlo method.}
\usage{
beta_ineq_sim(a, b, c, d, delta = 0, sims = 10000)
}
\arguments{
\item{a}{Parameter one of beta density for X}

\item{b}{Parameter two of beta density for X}

\item{c}{Parameter one of beta density for Y}

\item{d}{Parameter two of beta density for Y}

\item{delta}{The difference we wish to assess (i.e. X - Y > delta)}

\item{sims}{The number of Monte Carlo variates to generate for estimation}
}
\value{
The value of the integral
}
\description{
Calculate Pr(X > Y + delta) where X and Y are independent Beta random variables
using Monte Carlo method.
}
\examples{
beta_ineq_sim(5, 5, 3, 7)

}
