% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{calc_scenario_ppos}
\alias{calc_scenario_ppos}
\title{Calculate the Predictive Probability of Success for
a given trial scenario. Repeatedly calls calc_trial_ppos.}
\usage{
calc_scenario_ppos(scenario, useParallel = FALSE, ...)
}
\arguments{
\item{scenario}{A data.table returned from `sim_scenario`.}

\item{useParallel}{Use parallel processing for calculating PPoS.}

\item{...}{Other arguements to `calc_trial_ppos`.}
}
\value{
A data.table of scenario with the PPoS results incorporated.
}
\description{
Calculate the Predictive Probability of Success for
a given trial scenario. Repeatedly calls calc_trial_ppos.
}
