% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Calculate density of beta-binomial distribution}
\usage{
dbetabinom(x, n, a = 1, b = 1)
}
\arguments{
\item{x}{The value at which to evaluate the density}

\item{n}{The sample size}

\item{a}{First parameter}

\item{b}{Second parameter}
}
\value{
Value of beta-binomial(n,a,b) evaluated at x
}
\description{
Calculate density of beta-binomial distribution
}
\examples{
dbetabinom(5, 10, 2, 3)

}
