% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_stan.R
\name{lm_stan}
\alias{lm_stan}
\title{Logistic regression with Stan}
\usage{
lm_stan(X, y, n, ...)
}
\arguments{
\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{x}{Numeric vector of input values.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Logistic regression with Stan
}
