% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{plot_beta_norm}
\alias{plot_beta_norm}
\title{Compare Normal approximation to Beta}
\usage{
plot_beta_norm(a, b, ...)
}
\arguments{
\item{a}{First Beta parameter}

\item{b}{Second Beta parameter}

\item{...}{Other arguments to `curve()`}
}
\value{
A plot of Beta density and Normal approximation
}
\description{
Compare Normal approximation to Beta
}
