% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{plot_trial}
\alias{plot_trial}
\title{Plot a trial}
\usage{
plot_trial(trial, par_seq)
}
\arguments{
\item{trial}{A trial output from sim_trial}

\item{par_seq}{The sequence of parameter values at which to evaluate the densities}
}
\description{
Plot a trial
}
