% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{rbetabinom}
\alias{rbetabinom}
\title{Draw random variates from beta-binomial distribution}
\usage{
rbetabinom(n, m, a = 1, b = 1)
}
\arguments{
\item{n}{The number of random values to sample}

\item{m}{The sample size}

\item{a}{First parameter}

\item{b}{Second parameter}
}
\description{
Draw random variates from beta-binomial distribution
}
\examples{
rbetabinom(2, 10, 2, 3)

}
