% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{sim_trial}
\alias{sim_trial}
\title{Simulate Bayesian two-arm trial 
with early termination for futility/success
Model:
 p_1      ~ beta(a, b)
 p_2      ~ beta(c, d)
 x_i|p_i  ~ binomial(n_i, p_i), k = 1...K, i = 1,2
 p_i|x_i  ~ beta(a + sum(x_i), b + n_i - sum(x_i))}
\usage{
sim_trial(sim_id = 1, p1tru = 0.1, p2tru = 0.1, delta = 0,
  n1int = seq(100, 500, 50), n2int = seq(100, 500, 50), a1 = 1,
  b1 = 1, a2 = 1, b2 = 1, post_method = "exact")
}
\arguments{
\item{p1tru}{True value for response rate arm 1}

\item{p2tru}{True value for response rate arm 2}

\item{delta}{Relative difference of interest (X - Y > delta)}

\item{n1int}{Individuals with follow-up at each interim arm 1}

\item{n2int}{Individuals with follow-up at each interim arm 2}

\item{a1}{Prior parameter arm 1}

\item{b1}{Prior parameter arm 1}

\item{a2}{Prior parameter arm 2}

\item{b2}{Prior parameter arm 2}
}
\description{
Hypothesis:
  H_0: p_1 > p_2 + d
  H_1: p_1 < p_2 + d
}
\details{
Terimnal decision rule:
  Pr(p_1 < p_2 + d | x_i) > k_hi => H_1
  Pr(p_1 < p_2 + d | x_i) < k_lo => H_0
  Otherwise => inconclusive

Interim decision rule:
  Pr(p_1 < p_2 + d | x_i) > k_hi => H_1
  Pr(p_1 < p_2 + d | x_i) < k_lo => H_0

OR

PPoS > q_hi => expect success => H_1
  PPoS < q_lo => futile => H_0

Otherwise => continue to k + 1
}
