% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{sim_trial_dat}
\alias{sim_trial_dat}
\title{Simulate trial data using Poisson process for accrual rate}
\usage{
sim_trial_dat(sim_id = 1, p1tru = 0.1, p2tru = 0.1, nmax = 100,
  enro_rate = 1, enro_intensity = function(t) 1,
  resp_delay = function(n) runif(n))
}
\arguments{
\item{sim_id}{The id for the simulated trial data}

\item{p1tru}{True response rate under control}

\item{p2tru}{True response rate under treatment}

\item{nmax}{The maximum total sample size}

\item{enro_rate}{The baseline acrrual rate}

\item{enro_intensity}{Function for changing acrrual intensity}

\item{resp_delay}{Function returning response time value}
}
\description{
Simulate trial data using Poisson process for accrual rate
}
