% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{dec_trial}
\alias{dec_trial}
\title{Apply decision rule to trial data}
\usage{
dec_trial(trial, fut_k = 0.1, suc_k = 0.9, inf_k = 0.05,
  sup_k = 0.95)
}
\description{
Apply decision rule to trial data
}
