% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm_functions.R
\name{est_trial_prob}
\alias{est_trial_prob}
\title{Calculate trial probabilities (posterior and predictive)}
\usage{
est_trial_prob(d, ppos_q = 0.95, ppos_sim = 1000,
  post_method = "approx", a1 = 1, b1 = 1, a2 = 1, b2 = 1)
}
\arguments{
\item{d}{The Trial data}

\item{ppos_q}{The vector of cut-points to consider}

\item{ppos_sim}{Number of simulations from posterior predictive to use in PPoS(q) calculation}

\item{post_method}{Method used to calculate P(X > Y + delta), one of `exact` (integration), `approx` (normal), `sim` (monte carlo)}

\item{a1}{Prior shape for treatment 1}

\item{b1}{Prior scale for treatment 1}

\item{a2}{Prior shape for treatment 2}

\item{b2}{Prior scale for treatment 2}
}
\value{
Updates `d` inplace but also returns the updated `d`. 
`post` - current posterior probability, 
`post_int` - posterior probability when follow-up enrolled individuals,
`post_fin` - posterior probability when follow-up to maximum sample size,
`ppos_intq` - predictive probability of success given q if follow-up enrolled individuals,
`ppos_finq` - predictive probability of success given q if follow-up to maximum sample size,
}
\description{
Calculate trial probabilities (posterior and predictive)
}
